RST progress
Today I started coding my RST. First of all, I had to create the package
with the alien and player images, and the various classes I will need
such as the Player, Alien, SpaceInvaders, and Beam classes. I started
coding in the SpaceInvaders class by creating the menu screen. I have
the title, instructions, borders, and enter to start text displayed. I
have also made it so that they adjust when the size of the screen is
changed. Lastly, I have also set up some framework for the other methods
that I will need to implement later on. Next steps are to create the
display for the game and the transition from the menu screen to the game
screen (recognizing when the user presses enter).