RST progress
I have made a lottttt of progress today. First of all, I now detect when
the user has pressed enter. When they have, the menu text disappears and
the player is generated to the screen. I have created most of the
methods for the player class, including hide, show, and generatePlayer.
The user is able to move the player left or right using the left or
right arrows or A or D. If the player runs in to the left or right
border, it is moved back and not allowed to move any further. If the
screen is resized, the player will still not be able to go past the
borders. Also, the player will get pushed leftwards by the border if the
user resizes the screen to be smaller so that it is still on the screen.
A problem that I encountered was I thought the start method ran with
every iteration, so I planned to have different elements be added to the
root based on what phase of the game the player wants to access.
However, it only runs once, so I need to add all of the elements to the
root at the beginning. Since I had to do this, I also had to make
methods to hide or show certain elements. I also added a couple
variables, such as BORDER_WIDTH, PLAYER_WIDTH, and PLAYER_HEIGHT. The
last change to my design was that I changed the player data field of the
player class to playerImage for better comprehensibility (player.player
could be a little confusing). Next steps are to add in the beam class
for the beams, generate the beams, and the shooting mechanic.