RST Progress
I HAVE THE BEAM SHOOTING WORKING WOOOOOO. The best part is that it works
perfectly by following my design. How it works is that the beam objects
are first generated and added to the root node. There is a set amount of
beams that can be displayed on the screen at any time. They are set as
hidden at first so that they are not seen during the menu screen. Then,
if the up arrow is pressed, a for loop runs through the array of beam
objects and checks if any of them are equal to or above the top of the
screen. If they are, the beam is "shot", that is its x and y values are
set to be slightly above the player, and the index of the for loop is
set to the max so it ends. This re-uses old beams, but only if they are
at the top of the screen. Otherwise, if all of the beams are in the
process of being shot, another beam cannot be shot. If the beams are
below the top of the screen, they are moved upwards by the beam speed
every iteration. If they reach the top of the screen, they are set to
hidden and stop moving upwards. Some slight changes to the formatting of
my design are that I added the constants BEAM_WIDTH and BEAM_HEIGHT for
the generation of beams and placement of beams. Also, I created the
updateBeams() method to update the beams each frame instead of having
all of the code in the game timer. I also added show/hide methods in the
Beam class.