RST progress
Today, I have managed to get a row of aliens to spawn at the top of the
screen, continuously move down toward the player, re spawn back at the
top once they reach the player, and randomize their type again when they
are re spawned back at the top. When the game is complete, they will
only all re spawn back at the top if all of the aliens in that row are
vanquished. If an alive alien reaches the player, the user will have
lost and the loss screen will display. Methods that I had to create were
the randomizeType() method that takes the green alien chance and the red
alien chance and generates them randomly. If the sum of the chances are
less than the amount of aliens per row, the extra chances are allocated
to void aliens, which are just seen as an empty spot. When the type is
chosen, their respective point value, health, and vanquish status are
also set. Furthermore, I have the hide/show methods for when I want to
perform any of these actions to the aliens during the game. In the
SpaceInvaders class, I had to use a for loop to generate an alien object
for each spot in the first row array, then add them to the node to be
displayed. Upon level startup, they are positioned based on the y value
set for the top row, the width of the screen, as well as the spacing set
for between each alien.