RST progress
I spent most of the night developing the alien generation in rows. What
happens is on startup, they are all generated at the same spot. So, I
only allow them to move down or be visible if they are not intersecting
with the row that comes before it. After they have all separated, then I
no longer need to check if they are intersecting, and they just
continuously move down until they reach the bottom, at which case they
are randomized and re spawned at the top (this wont be like this for the
final product, its just for testing for now). Since I have multiple
rows, I made an 2D array that acts as an array for all of my row arrays.
This allows me to simplify my program by being able to use two for loops
instead of one separate one for every row I want to add. It also allows
me to change how many rows I want to have, since the for loop runs from
0 to the length of the 2D array. It allows my program to have more
flexibility. A problem that I encountered was the aliens showing
prematurely when I was using bounds to detect if they were intersecting.
It would say that they weren't intersecting at startup when they were,
so it would always get triggered to update once right at the beginning
which would give the appearance of them being shown right away. I solved
this problem by using a difference approach to detecting if one row is
intersecting with the other. I get the minimum y of the previous alien,
and see if it is greater or equal to the maximum y of the current alien.
Next steps are to implement collision detection between the beams and
the aliens.