RST progress
Today I developed the collision detection between the beams and the
aliens. When a beam is updated, it checks if it is intersecting with any
of the aliens on the screen that are alive. If it is intersecting, the
beam is set to be hidden, and the beam's damage is subtracted from the
alien's health. If the alien's health is equal to or less than 0, it is
vanquished, so it disappears from the screen and set to vanquished. I
had to add a data field to the beam class: the boolean "hit". The
program will only check for intersections if the beam's hit data field
is set to false. This prevents the same beam from intersecting with an
alien more than once and with more than one aliens.