RST progress
I have now added a loss screen that shows when the user loses the game.
That is, one of the aliens reach the bottom. The user is presented with
a button, which gives them the option to restart the game. If pressed,
this button will restart the level, with the aliens, player, and score
completely reset. I have started to add a boss that runs every time the
score reaches the boss sentinel, but the image is not displaying for
some reason. I'll have to look at my code a little bit more to try to
figure out what's going on.