RST progress
The first thing that I have fully added is the pause screen. When the
user presses the space bar, the pause screen appears and the game stops.
When the user presses the space bar again, the pause screen disappears
and the game resumes. The pause screen is only shown when the user
presses the pause button when the level is running. The second thing
that I have implemented is a boss. This boss is run when the score hits
a certain increment. I have the boss sentinel set to 200, so every 200
points, the boss is run. When the boss is run, the other aliens are
hidden and stop moving. The boss' movement is modeled by a cosine
function, and reverses direction every time it hits a border. It also
shoots a beam downwards at random intervals. When it does this, its
animation frame is changed until the beam reaches the bottom of the
screen. If the player intersects with the boss beam, they lose and the
loss screen is shown. The third thing that I added is difficulty
scaling. I count how many times the boss has been vanquished, and all of
the aliens' health and the boss alien health are multiplied by this
number.